+++ toy-rpc/Cargo.toml
@@ -29,7 +29,7 @@ rustdoc-args = ["--cfg", "feature=\"docs\""]
 [features]
 default = [
     "serde_bincode",
-    "async_std_runtime",
+    "tokio_runtime",
     "server",
     "client",
     "tls"
